+++ src/components/Lists.tsx
@@ -1,10 +1,18 @@
 import React, {Dispatch, SetStateAction, useState} from 'react';
 import {TodoData} from "../type/TodoData";
-import {DragDropContext, Draggable, Droppable, DropResult} from "react-beautiful-dnd";
+import {
+  DragDropContext,
+  Draggable,
+  DraggableProvided, DraggableStateSnapshot,
+  Droppable,
+  DroppableProvided,
+  DropResult
+} from "react-beautiful-dnd";
+import Item from "./Item";
 
 type Props = {
-  todoDatas: TodoData[]
-  setTodoDatas: Dispatch<SetStateAction<TodoData[]>>
+  todoDatas: TodoData[];
+  setTodoDatas: Dispatch<SetStateAction<TodoData[]>>;
 };
 
 export default function Lists(
@@ -19,50 +27,16 @@ export default function Lists(
     newTodoDatas.splice(result.destination.index, 0, reorderItem);
     setTodoDatas(newTodoDatas)
   };
-  const changeCheckbox = (id: number): void => {
-    const changeTodoDatas = todoDatas.map((item) => {
-      if (item.id === id){
-        item.completed = !item.completed;
-      }
-      return item;
-    });
-    setTodoDatas(changeTodoDatas);
-  }
-
-  const deleteItem = (id: number): void => {
-    const newTodoData = todoDatas.filter((data: TodoData) => data.id !== id)
-    setTodoDatas(newTodoData)
-  }
   return (
       <div>
         <DragDropContext onDragEnd={handleEnd}>
           <Droppable droppableId='handleEnd'>
-            {(provided) => (
+            {(provided:DroppableProvided) => (
                 <div {...provided.droppableProps} ref={provided.innerRef}>
                   {todoDatas.map((data:TodoData, index: number) => (
                     <Draggable draggableId={data.id.toString()} key={data.id} index={index}>
-                      {(provided, snapshot) => (
-                        <div key={data.id} {...provided.draggableProps} ref={provided.innerRef} {...provided.dragHandleProps} className={`${snapshot.isDragging ? "bg-gray-400": "bg-gray-100"} flex items-center justify-between w-full px-4 py-1 my-2 text-gray-600 border rounded`}>
-                          <div className="items-center">
-                            <input
-                              type="checkbox"
-                              defaultChecked={data.completed}
-                              onChange={() => changeCheckbox(data.id)}/>
-                            <span
-                                className={data.completed ? "line-through" : undefined}
-                            >
-                              {data.title}
-                            </span>
-                          </div>
-                          <div className="items-center">
-                            <button
-                                onClick={() => deleteItem(data.id)}
-                                className="px-4 py-2 float-right"
-                            >
-                              x
-                            </button>
-                          </div>
-                        </div>
+                      {(provided:DraggableProvided, snapshot:DraggableStateSnapshot) => (
+                        <Item data={data} provided={provided} snapshot={snapshot} todoDatas={todoDatas} setTodoDatas={setTodoDatas} />
                       )}
                     </Draggable>
                 ))}
